refactor: split decoration module and fix critical memory/concurrency issues
Split the 1062-line decoration.ts into three focused modules:
- blockParser.ts: standalone functions for markdown/rst/plaintext block parsing
- commentDecoration.ts: CommentDecoration and MarkdownDecoration render classes
- decoration.ts: CommentDecorationManager orchestration only (~440 lines)

fix(replaceSelections): bound variableCompletionMap to 50 entries with LRU
eviction; register workspace.onDidCloseTextDocument in extension.ts to purge
cache entries keyed by closed document URI

fix(manager): remove pendingTranslateTasks.clear() from config change handlers
Cache keys include the target language code, so in-flight tasks cannot collide
with new requests after a language switch. Clearing the pending map only caused
duplicate API calls without preventing stale data from being served.
