+++ src/languageFeature/decoration.ts
@@ -1,37 +1,26 @@
-import { compileBlock } from "../syntax/compile";
 import {
     window,
     Selection,
-    DecorationOptions,
-    TextEditorDecorationType,
     Disposable,
     TextDocument,
     workspace,
     Range,
     commands,
 } from "vscode";
-import { ctx } from "../extension";
 import { usePlaceholderCodeLensProvider } from "./codelen";
 import { getConfig, onConfigChange } from "../configuration";
-import { checkScopeFunction, ICommentBlock, ICommentToken, ITranslatedText } from "../interface";
+import { ICommentBlock } from "../interface";
 import { debounce } from "../util/short-live";
-import { getTextLength } from "../util/string";
-import { cachedTranslate, translateManager } from "../translate/manager";
 import { createComment } from "../syntax/Comment";
-
-interface IFenceState {
-    marker: '`' | '~';
-    len: number;
-}
-// helper interface for markdown code fence scanning result
-interface IMarkdownFenceScanResult {
-    occupiedLines: Set<number>;
-    inFenceAtRangeStart: boolean;
-}
-
-interface ITextBlockOptions {
-    isBoundary?: (trimText: string) => boolean;
-}
+import {
+    getExpandedVisibleRange,
+    scanMarkdownFenceLines,
+    getMarkdownTextBlocks,
+    isRstStructureBoundary,
+    getParagraphTextBlocks,
+    isMarkdownCodeScope,
+} from "./blockParser";
+import { CommentDecoration, MarkdownDecoration } from "./commentDecoration";
 
 class CommentDecorationManager {
     private static readonly HIDDEN_DOC_CACHE_LIMIT = 2;
@@ -225,248 +214,6 @@ class CommentDecorationManager {
         }, delay);
     }
 
-    private _parseFence(trimText: string): IFenceState | null {
-        const match = trimText.match(/^([`~]{3,})/);
-        if (!match || !match[1]) {
-            return null;
-        }
-
-        const markerText = match[1];
-        const marker = markerText[0];
-        if ((marker !== '`' && marker !== '~') || markerText.split('').some((item) => item !== marker)) {
-            return null;
-        }
-
-        return {
-            marker,
-            len: markerText.length
-        };
-    }
-
-    private _isFenceClose(trimText: string, fence: IFenceState): boolean {
-        const close = trimText.match(/^([`~]{3,})\s*$/);
-        if (!close || !close[1]) {
-            return false;
-        }
-
-        const markerText = close[1];
-        return markerText[0] === fence.marker && markerText.length >= fence.len;
-    }
-
-    private _getFenceStateBeforeLine(document: TextDocument, endLineExclusive: number): IFenceState | null {
-        let fence: IFenceState | null = null;
-
-        for (let line = 0; line < endLineExclusive; line++) {
-            const text = document.lineAt(line).text.trimStart();
-            if (!fence) {
-                const openFence = this._parseFence(text);
-                if (openFence) {
-                    fence = openFence;
-                }
-                continue;
-            }
-
-            if (this._isFenceClose(text, fence)) {
-                fence = null;
-            }
-        }
-
-        return fence;
-    }
-
-    private _isMarkdownCodeScope: checkScopeFunction = (scopes) => {
-        const rules = [
-            /^meta\.embedded\.block\.shellscript/,
-            /^variable\.other\.rust/,
-            /^meta\.embedded\.block\.rust/,
-            /^markup\.fenced_code\.block\.markdown/,
-            /^meta\.embedded\.block\./
-        ];
-
-        return scopes.some((scope) => rules.some((rule) => rule.test(scope)));
-    };
-
-    private _getExpandedVisibleRange(document: TextDocument, visibleRange: Range, paddingLines: number = 120) {
-        const startLine = Math.max(0, visibleRange.start.line - paddingLines);
-        const endLine = Math.min(document.lineCount - 1, visibleRange.end.line + paddingLines);
-        const endLineText = document.lineAt(endLine).text;
-        return new Range(startLine, 0, endLine, endLineText.length);
-    }
-
-    private _scanMarkdownFenceLines(document: TextDocument, visibleRange: Range): IMarkdownFenceScanResult {
-        let fence = this._getFenceStateBeforeLine(document, visibleRange.start.line);
-        const occupiedLines = new Set<number>();
-        const inFenceAtRangeStart = !!fence;
-
-        for (let line = visibleRange.start.line; line <= visibleRange.end.line; line++) {
-            const text = document.lineAt(line).text;
-            const trimText = text.trimStart();
-
-            if (!fence) {
-                const openFence = this._parseFence(trimText);
-                if (openFence) {
-                    occupiedLines.add(line);
-                    fence = openFence;
-                    continue;
-                }
-            } else {
-                occupiedLines.add(line);
-                if (this._isFenceClose(trimText, fence)) {
-                    fence = null;
-                }
-            }
-        }
-
-        return {
-            occupiedLines,
-            inFenceAtRangeStart
-        };
-    }
-
-    private _getMarkdownTextBlocks(document: TextDocument, visibleRange: Range, occupiedLineSet: Set<number>, inFenceAtRangeStart: boolean): ICommentBlock[] {
-        const blocks: ICommentBlock[] = [];
-        let paragraphStart = -1;
-        let paragraphTexts: string[] = [];
-
-        const pushParagraph = () => {
-            if (paragraphStart < 0 || paragraphTexts.length === 0) {
-                paragraphStart = -1;
-                paragraphTexts = [];
-                return;
-            }
-
-            const endLine = paragraphStart + paragraphTexts.length - 1;
-            const endLineText = document.lineAt(endLine).text;
-            blocks.push({
-                range: new Range(paragraphStart, 0, endLine, endLineText.length),
-                comment: paragraphTexts.join('\n')
-            });
-
-            paragraphStart = -1;
-            paragraphTexts = [];
-        };
-
-        let fence = inFenceAtRangeStart ? this._getFenceStateBeforeLine(document, visibleRange.start.line) : null;
-        for (let line = visibleRange.start.line; line <= visibleRange.end.line; line++) {
-            const text = document.lineAt(line).text;
-            const trimText = text.trimStart();
-
-            if (!fence) {
-                const openFence = this._parseFence(trimText);
-                if (openFence) {
-                    pushParagraph();
-                    fence = openFence;
-                    continue;
-                }
-            } else if (this._isFenceClose(trimText, fence)) {
-                pushParagraph();
-                fence = null;
-                continue;
-            }
-
-            if (fence || occupiedLineSet.has(line)) {
-                pushParagraph();
-                continue;
-            }
-
-            if (trimText.trim().length === 0) {
-                pushParagraph();
-                continue;
-            }
-
-            if (this._isMarkdownStructureBoundary(trimText)) {
-                pushParagraph();
-                continue;
-            }
-
-            if (paragraphStart < 0) {
-                paragraphStart = line;
-            }
-            paragraphTexts.push(text);
-        }
-
-        pushParagraph();
-
-        return blocks;
-    }
-
-    private _isMarkdownStructureBoundary(trimText: string): boolean {
-        if (/^(?:[-*_]\s*){3,}$/.test(trimText)) {
-            return true;
-        }
-
-        const tableSeparator = /^\|?\s*:?-{3,}:?\s*(?:\|\s*:?-{3,}:?\s*)+\|?$/;
-        return tableSeparator.test(trimText);
-    }
-
-    private _isRstStructureBoundary(trimText: string): boolean {
-        if (!trimText) {
-            return false;
-        }
-
-        if (/^([-=*~`^"'+#])\1{2,}$/.test(trimText)) {
-            return true;
-        }
-
-        if (/^\.\.\s+/.test(trimText)) {
-            return true;
-        }
-
-        if (/^[:\w-]+:\s+/.test(trimText)) {
-            return true;
-        }
-
-        return false;
-    }
-
-    private _getParagraphTextBlocks(document: TextDocument, range: Range, options: ITextBlockOptions = {}): ICommentBlock[] {
-        const blocks: ICommentBlock[] = [];
-        let paragraphStart = -1;
-        let paragraphTexts: string[] = [];
-        const isBoundary = options.isBoundary;
-
-        const pushParagraph = () => {
-            if (paragraphStart < 0 || paragraphTexts.length === 0) {
-                paragraphStart = -1;
-                paragraphTexts = [];
-                return;
-            }
-
-            const endLine = paragraphStart + paragraphTexts.length - 1;
-            const endLineText = document.lineAt(endLine).text;
-            blocks.push({
-                range: new Range(paragraphStart, 0, endLine, endLineText.length),
-                comment: paragraphTexts.join('\n')
-            });
-
-            paragraphStart = -1;
-            paragraphTexts = [];
-        };
-
-        for (let line = range.start.line; line <= range.end.line; line++) {
-            const text = document.lineAt(line).text;
-            const trimText = text.trimStart();
-
-            if (trimText.trim().length === 0) {
-                pushParagraph();
-                continue;
-            }
-
-            if (isBoundary && isBoundary(trimText)) {
-                pushParagraph();
-                continue;
-            }
-
-            if (paragraphStart < 0) {
-                paragraphStart = line;
-            }
-            paragraphTexts.push(text);
-        }
-
-        pushParagraph();
-        return blocks;
-    }
-
     private updateAvgRenderCost(cost: number) {
         if (cost <= 0) {
             return;
@@ -503,9 +250,9 @@ class CommentDecorationManager {
         try {
             const visibleRange = editor.visibleRanges[0];
             if (this.currDocument.languageId === 'markdown') {
-                const parseRange = this._getExpandedVisibleRange(this.currDocument, visibleRange);
+                const parseRange = getExpandedVisibleRange(this.currDocument, visibleRange);
                 const occupiedLineSet = new Set<number>();
-                const fenceScan = this._scanMarkdownFenceLines(this.currDocument, parseRange);
+                const fenceScan = scanMarkdownFenceLines(this.currDocument, parseRange);
                 fenceScan.occupiedLines.forEach((line) => occupiedLineSet.add(line));
 
                 const scopeFallback = getConfig<boolean>('markdown.scopeFallback', true);
@@ -514,7 +261,7 @@ class CommentDecorationManager {
                         const comment = await createComment();
                         const codeBlocks = await comment.getAllScopeBlocks(
                             this.currDocument,
-                            this._isMarkdownCodeScope,
+                            isMarkdownCodeScope,
                             parseRange
                         ) || [];
                         codeBlocks.forEach((block) => {
@@ -533,7 +280,7 @@ class CommentDecorationManager {
                     }
                 }
 
-                blocks = this._getMarkdownTextBlocks(
+                blocks = getMarkdownTextBlocks(
                     this.currDocument,
                     parseRange,
                     occupiedLineSet,
@@ -542,13 +289,13 @@ class CommentDecorationManager {
 
                 blocks = blocks.filter((block) => !!block.range.intersection(visibleRange));
             } else if (this.currDocument.languageId === 'restructuredtext' || this.currDocument.languageId === 'rst') {
-                const parseRange = this._getExpandedVisibleRange(this.currDocument, visibleRange);
-                blocks = this._getParagraphTextBlocks(this.currDocument, parseRange, {
-                    isBoundary: this._isRstStructureBoundary.bind(this)
+                const parseRange = getExpandedVisibleRange(this.currDocument, visibleRange);
+                blocks = getParagraphTextBlocks(this.currDocument, parseRange, {
+                    isBoundary: isRstStructureBoundary
                 }).filter((block) => !!block.range.intersection(visibleRange));
             } else if (this.currDocument.languageId === 'plaintext' || this.currDocument.languageId === 'text') {
-                const parseRange = this._getExpandedVisibleRange(this.currDocument, visibleRange);
-                blocks = this._getParagraphTextBlocks(this.currDocument, parseRange)
+                const parseRange = getExpandedVisibleRange(this.currDocument, visibleRange);
+                blocks = getParagraphTextBlocks(this.currDocument, parseRange)
                     .filter((block) => !!block.range.intersection(visibleRange));
             } else {
                 let comment = await createComment();
@@ -582,7 +329,7 @@ class CommentDecorationManager {
             if (this.currDocument!.languageId === 'markdown' && (!block.tokens || block.tokens.length === 0)) {
                 commentDecoration = new MarkdownDecoration(block, this.currDocument!, this.inplace);
             } else {
-                commentDecoration = new CommentDecoration(block, this.currDocument!, this.inplace)
+                commentDecoration = new CommentDecoration(block, this.currDocument!, this.inplace);
             }
 
             newBlockMaps.set(key, { comment: block.comment, commentDecoration });
@@ -685,377 +432,6 @@ class CommentDecorationManager {
         cleanAll();
         this.currDocument = undefined;
         this.showBrowseCommentTranslateImpl();
-    }
-}
-
-class CommentDecoration {
-    private _loading: boolean = true;
-    private _desposed: boolean = false;
-    private _loadingDecoration: TextEditorDecorationType;
-    private _translatedDecoration: TextEditorDecorationType | undefined;
-    private _contentDecorations: DecorationOptions[] = [];
-
-    constructor(protected _block: ICommentBlock, private _currDocument: TextDocument, private _inplace: boolean = false) {
-        this._loadingDecoration = window.createTextEditorDecorationType({
-            after: {
-                contentIconPath: ctx.asAbsolutePath("resources/icons/loading.svg"),
-            },
-        });
-        this.reflash();
-        this.translate();
-    }
-
-    get block() {
-        return this._block;
-    }
-
-    editing(): boolean {
-        let range = this._block.range;
-        let selection = window.activeTextEditor?.selection;
-        if (selection) {
-            return range.intersection(selection) ? true : false;
-        }
-        return false;
-    }
-
-    private _getRstPrefixLen(line: string): number {
-        let remaining = line;
-        let consumed = 0;
-
-        const leadingWhitespace = remaining.match(/^\s*/)?.[0] || '';
-        consumed += leadingWhitespace.length;
-        remaining = remaining.slice(leadingWhitespace.length);
-
-        const bulletList = remaining.match(/^[-*+]\s+/);
-        if (bulletList) {
-            return consumed + bulletList[0].length;
-        }
-
-        const orderedList = remaining.match(/^\d+[.)]\s+/);
-        if (orderedList) {
-            return consumed + orderedList[0].length;
-        }
-
-        return consumed;
-    }
-
-    protected async _compile(): Promise<ITranslatedText | null> {
-        let { tokens } = this._block;
-        if (!tokens || tokens.length === 0) {
-            const lines = this._block.comment.split('\n');
-            const isRst = this._currDocument.languageId === 'restructuredtext' || this._currDocument.languageId === 'rst';
-            const fallbackTokens: ICommentToken[] = lines.map((line) => {
-                const ignoreStart = isRst
-                    ? this._getRstPrefixLen(line)
-                    : (line.match(/^\s*/)?.[0] || '').length;
-                return {
-                    text: line,
-                    ignoreStart,
-                    ignoreEnd: 0,
-                };
-            });
-
-            const texts = fallbackTokens.map(({ text, ignoreStart = 0, ignoreEnd = 0 }) => {
-                return text.slice(ignoreStart, text.length - ignoreEnd).trim();
-            });
-
-            const sourceText = texts.join('\n');
-            const translatedText = sourceText.length > 0
-                ? await cachedTranslate(sourceText, { to: translateManager.opts.to || 'en' })
-                : '';
-
-            const targetLines = translatedText.split('\n');
-            let targets: string[];
-            if (targetLines.length === lines.length) {
-                targets = targetLines;
-            } else if (targetLines.length === 1) {
-                targets = [targetLines[0], ...new Array(Math.max(lines.length - 1, 0)).fill('')];
-            } else {
-                const merged = targetLines.join(' ').trim();
-                targets = [merged, ...new Array(Math.max(lines.length - 1, 0)).fill('')];
-            }
-
-            this._block.tokens = fallbackTokens;
-
-            return {
-                translatedText,
-                targets,
-                texts,
-                combined: [],
-                translateLink: ''
-            };
-        }
-        return compileBlock(this._block, this._currDocument.languageId);
-    }
-
-    // Translate commentblock text and set decorative content
-    async translate() {
-        let result: ITranslatedText | null = null;
-        try {
-            result = await this._compile();
-        } catch (error) {
-            console.error('[comment-translate] translate compile failed:', error);
-        }
-
-        let { tokens, range } = this._block;
-        this._loading = false;
-
-        if (!result || !tokens || tokens.length === 0) {
-            this.reflash();
-            return null;
-        }
-
-        let { targets, texts, combined, translatedText } = result;
-
-        let targetIndex = 0;
-        let tokensLength = tokens.length || 0;
-        tokens.forEach((token, i) => {
-            let { text, ignoreStart = 0, ignoreEnd = 0 } = token;
-            const translateText = texts[i];
-            let targetText = translateText.length > 0 ? targets[targetIndex++] : "";
-            let offset = i === 0 ? range.start.character : 0;
-            let originText = text.slice(ignoreStart, text.length - ignoreEnd);
-
-            let combinedIndex = i;
-
-            // No need to display decoration when translation is wrong
-            if (this._inplace && translatedText) {
-                this._contentDecorations.push({
-                    range: new Selection(
-                        range.start.line + combinedIndex,
-                        offset + ignoreStart,
-                        range.start.line + combinedIndex,
-                        offset + text.length - ignoreEnd
-                    ),
-                    renderOptions: this.genrateDecorationOptions(targetText),
-                });
-                return;
-            }
-
-            for (let k = i + 1; k < tokensLength && combined[k]; k++) {
-                combinedIndex = k;
-            }
-
-            if (targetText && targetText !== originText) {
-                // Read text here will be more than blocks and need to be read throughout the document
-                const showLine = Math.min(this._currDocument.lineCount - 1, range.start.line + combinedIndex + 1);
-                let showLineLen = getTextLength(this._currDocument!.lineAt(showLine).text);
-                // let offsetText = texts[combinedIndex].substring(0, offset + ignoreStart);
-                let offsetText = this._currDocument!.lineAt(range.start.line + combinedIndex).text.substring(0, offset + ignoreStart);
-
-                let gap = getTextLength(offsetText) - showLineLen;
-                if (gap > 0) {
-                    targetText = targetText.padStart(targetText.length + gap, '\u00a0');
-                }
-                this._contentDecorations.push({
-                    range: new Selection(
-                        range.start.line + combinedIndex + 1,
-                        offset + ignoreStart,
-                        range.start.line + combinedIndex + 1,
-                        offset + text.length - ignoreEnd
-                    ),
-                    renderOptions: this.genrateDecorationOptions(targetText),
-                });
-            }
-        });
-
-        this.reflash();
-    }
-
-    // Generate the text decoration styles
-    genrateDecorationOptions(text: string) {
-        if (this._inplace) {
-            return {
-                before: {
-                    color: `var(--vscode-editorCodeLens-foreground)`,
-                    // textDecoration: text.trim().length > 0 ? `none;word-wrap: break-word; white-space: pre-wrap;display: inline-block; width:calc(${showLineLen}ch); max-height: ${lines}lh; position: relative;` : '',
-                    contentText: text,
-                },
-            };
-        }
-        return {
-            before: {
-                textDecoration: `none; font-size: 1em; display: inline-block; position: relative; width: 0; top: ${-1.3}em;`,
-                contentText: text,
-                color: 'var(--vscode-editorCodeLens-foreground)',
-            },
-        }
-    }
-
-    // Get the text decoration type
-    getTranslatedDecoration() {
-        if (this._translatedDecoration) {
-            return this._translatedDecoration;
-        }
-
-        let textDecoration: string | undefined;
-        if (this._inplace) {
-            textDecoration = "none; display: none;";
-        }
-        this._translatedDecoration = window.createTextEditorDecorationType({
-            textDecoration,
-        });
-
-        return this._translatedDecoration;
-    }
-
-    // 重新渲染装饰内容
-    reflash() {
-        if (this._desposed) return;
-        const currentUri = this._currDocument.uri.toString();
-        const targetEditors = window.visibleTextEditors.filter((editor) => {
-            return editor.document.uri.toString() === currentUri;
-        });
-
-        if (targetEditors.length === 0) {
-            return;
-        }
-
-        if (this._loading) {
-            targetEditors.forEach((editor) => {
-                editor.setDecorations(this._loadingDecoration, [
-                    this._block.range,
-                ]);
-            });
-            return;
-        } else {
-            targetEditors.forEach((editor) => {
-                editor.setDecorations(this._loadingDecoration, []);
-            });
-        }
-
-        let translatedDecoration = this.getTranslatedDecoration();
-        // The current comment block is being edited, translation status is not displayed
-        if (this._inplace && this.editing()) {
-            targetEditors.forEach((editor) => {
-                editor.setDecorations(translatedDecoration, []);
-            });
-            return;
-        }
-
-        if (!this._inplace) {
-            let { append } = usePlaceholderCodeLensProvider();
-            let lines = this._contentDecorations.map((decoration) => decoration.range.start.line);
-            append(this._currDocument, lines);
-        }
-        targetEditors.forEach((editor) => {
-            editor.setDecorations(translatedDecoration, this._contentDecorations);
-        });
-    }
-
-    dispose() {
-        this._desposed = true;
-        this._loadingDecoration.dispose();
-        this._translatedDecoration?.dispose();
-    }
-}
-
-
-class MarkdownDecoration extends CommentDecoration {
-    constructor(block: ICommentBlock, currDocument: TextDocument, inplace: boolean = false) {
-        super(block, currDocument, inplace);
-    }
-
-    private _getMarkdownPrefixLen(line: string): number {
-        let remaining = line;
-        let consumed = 0;
-
-        const leadingSpace = remaining.match(/^\s{0,3}/);
-        if (leadingSpace) {
-            consumed += leadingSpace[0].length;
-            remaining = remaining.slice(leadingSpace[0].length);
-        }
-
-        while (remaining.startsWith('>')) {
-            const quote = remaining.match(/^>\s?/);
-            if (!quote) {
-                break;
-            }
-            consumed += quote[0].length;
-            remaining = remaining.slice(quote[0].length);
-        }
-
-        const heading = remaining.match(/^#{1,6}\s+/);
-        if (heading) {
-            return consumed + heading[0].length;
-        }
-
-        const listTask = remaining.match(/^[-*+]\s+\[[ xX]\]\s+/);
-        if (listTask) {
-            return consumed + listTask[0].length;
-        }
-
-        const list = remaining.match(/^[-*+]\s+/);
-        if (list) {
-            return consumed + list[0].length;
-        }
-
-        const orderedList = remaining.match(/^\d+[.)]\s+/);
-        if (orderedList) {
-            return consumed + orderedList[0].length;
-        }
-
-        return consumed;
-    }
-
-    private _normalizeTargets(targetLines: string[], lineCount: number): string[] {
-        if (targetLines.length === lineCount) {
-            return targetLines;
-        }
-
-        if (targetLines.length === 1) {
-            return [targetLines[0], ...new Array(lineCount - 1).fill('')];
-        }
-
-        const merged = targetLines.join(' ').trim();
-        return [merged, ...new Array(lineCount - 1).fill('')];
-    }
-
-    protected async _compile(): Promise<ITranslatedText | null> {
-
-        let result: ITranslatedText;
-
-        let block = this._block;
-        const lines = block.comment.split('\n');
-        const tokens: ICommentToken[] = lines.map((line) => {
-            const ignoreStart = this._getMarkdownPrefixLen(line);
-            return {
-                text: line,
-                ignoreStart,
-                ignoreEnd: 0,
-            };
-        });
-
-        const texts = tokens.map(({ text, ignoreStart = 0, ignoreEnd = 0 }) => {
-            return text.slice(ignoreStart, text.length - ignoreEnd).trim();
-        });
-
-        const sourceText = texts.join('\n');
-        let translatedText = sourceText.length > 0
-            ? await cachedTranslate(sourceText, { to: translateManager.opts.to || 'en' })
-            : '';
-        const targets = this._normalizeTargets(translatedText.split('\n'), lines.length);
-
-        const reconstructedLines = tokens.map((token, index) => {
-            const { text, ignoreStart = 0, ignoreEnd = 0 } = token;
-            const startText = text.slice(0, ignoreStart);
-            const endText = text.slice(text.length - ignoreEnd);
-            const target = targets[index] || '';
-            return startText + target + endText;
-        });
-
-        translatedText = reconstructedLines.join('\n');
-        block.tokens = tokens;
-
-        result = {
-            translatedText,
-            targets,
-            texts,
-            combined: [],
-            translateLink: ''
-        }
-
-        return result;
     }
 }
